06/src/requirements.txt
#ex00
grpcio-tools
grpcio
protobuf
#ex01
pydantic
#ex02
sqlalchemy
psycopg2
